requirements.txt
fastapi
uvicorn
python-dotenv
openai
python-jose
python-multipart
aiofiles
